requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.25.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
bleak==0.21.1
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
cryptography==41.0.7
databases==0.4.2
dbus-fast==2.21.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
fastapi-mail==1.4.1
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
frozenlist==1.4.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
interchange==2021.0.4
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
locust==2.23.1
makefun==1.15.2
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
msgpack==1.0.7
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
neo4j==5.15.0
orjson==3.9.10
packaging==23.2
pandocfilters==1.5.1
pansi==2020.7.3
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py2neo==2021.2.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.33.0
requests==2.31.0
roundrobin==0.0.4
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.36.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
